requirements.txt
appnope==0.1.4
asttokens==2.4.1
bidict==0.23.1
blinker==1.9.0
click==8.1.7
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
dnspython==2.7.0
eventlet==0.38.0
executing==2.1.0
Flask==3.1.0
Flask-SocketIO==5.4.1
greenlet==3.1.1
h11==0.14.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mido==1.3.3
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
push2-python @ git+https://github.com/ffont/push2-python@ae91501788469f2c54aa95ed12f37f83505d73f8
pycairo==1.27.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-osc==1.9.0
python-rtmidi==1.5.8
python-socketio==5.11.4
pyusb==1.2.1
pyzmq==26.2.0
simple-websocket==1.1.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
